Updated Chapter 9 Exercises

+++ Chapter09/Exercise09.01/app/src/main/java/com/android/testable/myapplication/NumberAdder.kt
@@ -9,7 +9,7 @@ class NumberAdder {
         if (n < 0) {
             throw InvalidNumberException
         }
-        callback.invoke(n.toBigInteger().times((n.toLong() + 1).toBigInteger()).divide(2.toBigInteger()))
+        callback.invoke(n.toBigInteger().times((n.toBigInteger() + 1.toBigInteger())).divide(2.toBigInteger()))
     }
 
     object InvalidNumberException : Throwable()
